backend/requirements.txt
# ============================================================================
# BCC Planning Validation System - MVP (Runtime)
# Python 3.11+ required
# ============================================================================

# ----------------------------
# Core API Framework
# ----------------------------
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
slowapi==0.1.9  # Rate limiting for FastAPI

pydantic==2.5.3
pydantic-settings==2.1.0

# ----------------------------
# Database (PostgreSQL + PostGIS)
# ----------------------------
# Psycopg 3 (recommended for new projects)
# Using binary package for easier installation (includes libpq)
psycopg[binary]==3.3.2
psycopg-binary==3.3.2

SQLAlchemy==2.0.25
alembic==1.13.1
geoalchemy2==0.14.3
shapely==2.0.2

# ----------------------------
# Azure - Auth & Storage
# ----------------------------
azure-identity==1.15.0
msal==1.26.0
azure-storage-blob==12.19.0

# ----------------------------
# Azure - Secrets & Monitoring
# ----------------------------
azure-keyvault-secrets==4.7.0
opencensus-ext-azure==1.1.13

# ----------------------------
# Azure - AI Services
# ----------------------------
openai==1.10.0
azure-ai-documentintelligence==1.0.2

# ----------------------------
# Document Processing (OCR)
# ----------------------------
pdfplumber==0.10.3
pytesseract==0.3.10
Pillow==10.2.0
pdf2image==1.17.0
reportlab==4.2.5
PyMuPDF==1.23.8

# ----------------------------
# NLP (lightweight field extraction)
# ----------------------------
spacy==3.7.2
# NOTE: model install is separate:
#   python -m spacy download en_core_web_sm

# ----------------------------
# Web UI (HITL)
# ----------------------------
streamlit==1.30.0
streamlit-aggrid==0.3.4.post3
plotly==5.18.0

# ----------------------------
# HTTP Clients
# ----------------------------
httpx==0.26.0
requests==2.31.0

# ----------------------------
# Utilities
# ----------------------------
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
click==8.1.7  # CLI tool framework
fuzzywuzzy==0.18.0  # Fuzzy string matching for address comparison
python-Levenshtein==0.25.0  # Faster string matching for fuzzywuzzy

# ----------------------------
# Testing
# ----------------------------
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.6.0

# ----------------------------
# Code Quality
# ----------------------------
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# ----------------------------
# Logging & Monitoring
# ----------------------------
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0
psutil==5.9.8  # System metrics

# ----------------------------
# Optional: Background Jobs (enable when needed)
# ----------------------------
# rq==1.16.0
# redis==5.0.1
# celery==5.3.6
